feat: add customary volume units and enhance existing unit conversions
- Introduced new volume units for US customary measurements including Cubic Inch, Cubic Foot, US Gallon, and US Fluid Ounce.
- Updated the volume module to include these new units and their conversions.
- Enhanced the velocity module to conditionally include astronomical units based on feature flags.
- Added tests for new volume units and conversions, ensuring accuracy and consistency.
- Updated Cargo.toml files to enable all-units feature for qtty-ffi and qtty.
- Refactored scalar aliases to include new units and ensure compatibility with feature flags.
