+++ qtty-core/src/units/time/mod.rs
@@ -50,9 +50,6 @@
 //! touch!(Weeks,        1.0); touch!(Fortnights,   1.0);
 //! touch!(Years,        1.0); touch!(Decades,      1.0);
 //! touch!(Centuries,    1.0); touch!(Millennia,    1.0);
-//! touch!(JulianYears,  1.0); touch!(JulianCenturies, 1.0);
-//! touch!(SiderealDays, 1.0); touch!(SynodicMonths, 1.0);
-//! touch!(SiderealYears, 1.0);
 //! ```
 
 use crate::{Quantity, Unit};
@@ -67,6 +64,15 @@ impl<T: Unit<Dim = Time>> TimeUnit for T {}
 
 /// Conventional civil mapping used by this module: seconds per mean solar day.
 pub const SECONDS_PER_DAY: f64 = 86_400.0;
+
+#[cfg(feature = "julian-time")]
+mod julian_time;
+#[cfg(feature = "julian-time")]
+pub use julian_time::*;
+#[cfg(feature = "astro")]
+mod astro;
+#[cfg(feature = "astro")]
+pub use astro::*;
 
 // --- SI submultiples of the second ---
 
@@ -292,61 +298,6 @@ pub type Millennia = Quantity<Millennium>;
 /// A constant representing one millennium.
 pub const MILLENNIUM: Millennia = Millennia::new(1.0);
 
-// --- Julian conventions (useful in astronomy/ephemerides) ---
-
-/// Julian year (`365.25 d`), expressed in seconds.
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Unit)]
-#[unit(symbol = "a", dimension = Time, ratio = 365.25 * SECONDS_PER_DAY)]
-pub struct JulianYear;
-/// A quantity measured in Julian years.
-pub type JulianYears = Quantity<JulianYear>;
-/// A constant representing one Julian year.
-pub const JULIAN_YEAR: JulianYears = JulianYears::new(1.0);
-
-/// Julian century (`36_525 d`), expressed in seconds.
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Unit)]
-#[unit(symbol = "JC", dimension = Time, ratio = 36_525.0 * SECONDS_PER_DAY)]
-pub struct JulianCentury;
-/// A quantity measured in Julian centuries.
-pub type JulianCenturies = Quantity<JulianCentury>;
-/// A constant representing one Julian century.
-pub const JULIAN_CENTURY: JulianCenturies = JulianCenturies::new(1.0);
-
-// --- Astronomical mean units (explicitly approximate) ---
-
-/// Mean sidereal day (Earth), expressed in SI seconds.
-///
-/// Convention used: `1 sidereal day ≈ 86_164.0905 s`.
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Unit)]
-#[unit(symbol = "sd", dimension = Time, ratio = 86_164.090_5)]
-pub struct SiderealDay;
-/// A quantity measured in sidereal days.
-pub type SiderealDays = Quantity<SiderealDay>;
-/// A constant representing one sidereal day.
-pub const SIDEREAL_DAY: SiderealDays = SiderealDays::new(1.0);
-
-/// Mean synodic month (lunar phase cycle), expressed in seconds.
-///
-/// Convention used: `1 synodic month ≈ 29.530588 d`.
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Unit)]
-#[unit(symbol = "synmo", dimension = Time, ratio = 29.530_588 * SECONDS_PER_DAY)]
-pub struct SynodicMonth;
-/// A quantity measured in synodic months.
-pub type SynodicMonths = Quantity<SynodicMonth>;
-/// A constant representing one synodic month.
-pub const SYNODIC_MONTH: SynodicMonths = SynodicMonths::new(1.0);
-
-/// Mean sidereal year (Earth), expressed in seconds.
-///
-/// Common convention: `1 sidereal year ≈ 365.256363004 d`.
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Unit)]
-#[unit(symbol = "syr", dimension = Time, ratio = 365.256_363_004 * SECONDS_PER_DAY)]
-pub struct SiderealYear;
-/// A quantity measured in sidereal years.
-pub type SiderealYears = Quantity<SiderealYear>;
-/// A constant representing one sidereal year.
-pub const SIDEREAL_YEAR: SiderealYears = SiderealYears::new(1.0);
-
 #[cfg(all(test, feature = "std"))]
 mod tests {
     use super::*;
@@ -399,6 +350,7 @@ mod tests {
         assert_abs_diff_eq!(week.value(), 1.0, epsilon = 1e-12);
     }
 
+    #[cfg(feature = "julian-time")]
     #[test]
     fn julian_year_to_days() {
         let jy = JulianYears::new(1.0);
@@ -406,6 +358,7 @@ mod tests {
         assert_abs_diff_eq!(day.value(), 365.25, epsilon = 1e-9);
     }
 
+    #[cfg(feature = "julian-time")]
     #[test]
     fn julian_century_to_days() {
         let jc = JulianCenturies::new(1.0);
@@ -413,6 +366,7 @@ mod tests {
         assert_abs_diff_eq!(day.value(), 36525.0, epsilon = 1e-9);
     }
 
+    #[cfg(feature = "julian-time")]
     #[test]
     fn julian_century_to_julian_years() {
         let jc = JulianCenturies::new(1.0);
@@ -453,6 +407,7 @@ mod tests {
         assert_abs_diff_eq!(back.value(), original.value(), epsilon = 1e-12);
     }
 
+    #[cfg(feature = "julian-time")]
     #[test]
     fn roundtrip_julian_year_day() {
         let original = JulianYears::new(2.5);
@@ -504,6 +459,10 @@ mod tests {
             prop_assert!((sec.value() / day.value() - 86400.0).abs() < 1e-9);
         }
 
+    }
+
+    #[cfg(feature = "julian-time")]
+    proptest! {
         #[test]
         fn prop_julian_year_day_ratio(y in 1e-6..1e6f64) {
             let jy = JulianYears::new(y);
@@ -633,6 +592,7 @@ mod tests {
 
     // ─── Astronomical mean units ────────────────────────────────────────────
 
+    #[cfg(feature = "astro")]
     #[test]
     fn sidereal_day_to_seconds() {
         let q = SiderealDays::new(1.0);
@@ -641,6 +601,7 @@ mod tests {
         assert_abs_diff_eq!(s.value(), 86_164.090_5, epsilon = 1e-3);
     }
 
+    #[cfg(feature = "astro")]
     #[test]
     fn synodic_month_to_days() {
         let q = SynodicMonths::new(1.0);
@@ -649,6 +610,7 @@ mod tests {
         assert_abs_diff_eq!(d.value(), 29.530_588, epsilon = 1e-6);
     }
 
+    #[cfg(feature = "astro")]
     #[test]
     fn sidereal_year_to_days() {
         let q = SiderealYears::new(1.0);
@@ -665,6 +627,11 @@ mod tests {
         assert_eq!(format!("{}", Nanoseconds::new(1.0)), "1 ns");
         assert_eq!(format!("{}", Kiloseconds::new(1.0)), "1 ks");
         assert_eq!(format!("{}", Fortnights::new(1.0)), "1 fn");
+    }
+
+    #[cfg(feature = "astro")]
+    #[test]
+    fn astro_symbols_are_correct() {
         assert_eq!(format!("{}", SiderealDays::new(1.0)), "1 sd");
     }
 }
