+++ qtty/src/lib.rs
@@ -202,50 +202,88 @@ pub use qtty_core::units::volume;
 pub mod unit {
     pub use qtty_core::{Per, Prod, Unit, UnitDiv, UnitMul, Unitless};
 
+    #[cfg(feature = "navigation")]
+    pub use qtty_core::units::angular::Gradian;
+    pub use qtty_core::units::angular::{AngularUnit, Degree, Milliradian, Radian, Turn};
+    #[cfg(feature = "astro")]
     pub use qtty_core::units::angular::{
-        AngularUnit, Arcminute, Arcsecond, Degree, Gradian, HourAngle, MicroArcsecond,
-        MilliArcsecond, Milliradian, Radian, Turn,
+        Arcminute, Arcsecond, HourAngle, MicroArcsecond, MilliArcsecond,
     };
+
+    #[cfg(feature = "land-area")]
+    pub use qtty_core::units::area::{Acre, Are, Hectare};
     pub use qtty_core::units::area::{
-        Acre, Are, AreaUnit, Hectare, SquareCentimeter, SquareFoot, SquareInch, SquareKilometer,
-        SquareMeter, SquareMile, SquareMillimeter, SquareYard,
+        AreaUnit, SquareCentimeter, SquareKilometer, SquareMeter, SquareMillimeter,
     };
+    #[cfg(feature = "customary")]
+    pub use qtty_core::units::area::{SquareFoot, SquareInch, SquareMile, SquareYard};
+
+    #[cfg(feature = "astro")]
     pub use qtty_core::units::length::nominal::{
         EarthEquatorialRadius, EarthPolarRadius, EarthRadius, JupiterRadius, LunarDistance,
         LunarRadius, SolarDiameter, SolarRadius,
     };
+    #[cfg(feature = "astro")]
     pub use qtty_core::units::length::{
-        AstronomicalUnit, Attometer, BohrRadius, Centimeter, Chain, ClassicalElectronRadius,
-        Decameter, Decimeter, EarthEquatorialCircumference, EarthMeridionalCircumference,
-        ElectronReducedComptonWavelength, Exameter, Fathom, Femtometer, Foot, Gigameter,
-        Gigaparsec, Hectometer, Inch, Kilometer, Kiloparsec, LengthUnit, LightYear, Link,
-        Megameter, Megaparsec, Meter, Micrometer, Mile, Millimeter, Nanometer, NauticalMile,
-        Parsec, Petameter, Picometer, PlanckLength, Rod, Terameter, Yard, Yoctometer, Yottameter,
-        Zeptometer, Zettameter,
+        AstronomicalUnit, Gigaparsec, Kiloparsec, LightYear, Megaparsec, Parsec,
     };
+    pub use qtty_core::units::length::{
+        Attometer, Centimeter, Decameter, Decimeter, Exameter, Femtometer, Gigameter, Hectometer,
+        Kilometer, LengthUnit, Megameter, Meter, Micrometer, Millimeter, Nanometer, Petameter,
+        Picometer, Terameter, Yoctometer, Yottameter, Zeptometer, Zettameter,
+    };
+    #[cfg(feature = "fundamental-physics")]
+    pub use qtty_core::units::length::{
+        BohrRadius, ClassicalElectronRadius, ElectronReducedComptonWavelength, PlanckLength,
+    };
+    #[cfg(feature = "navigation")]
+    pub use qtty_core::units::length::{
+        Chain, EarthEquatorialCircumference, EarthMeridionalCircumference, Fathom, Link,
+        NauticalMile, Rod,
+    };
+    #[cfg(feature = "customary")]
+    pub use qtty_core::units::length::{Foot, Inch, Mile, Yard};
+
+    #[cfg(feature = "fundamental-physics")]
+    pub use qtty_core::units::mass::AtomicMassUnit;
+    #[cfg(feature = "astro")]
+    pub use qtty_core::units::mass::SolarMass;
     pub use qtty_core::units::mass::{
-        AtomicMassUnit, Attogram, Carat, Centigram, Decagram, Decigram, Exagram, Femtogram,
-        Gigagram, Grain, Gram, Hectogram, Kilogram, LongTon, MassUnit, Megagram, Microgram,
-        Milligram, Nanogram, Ounce, Petagram, Picogram, Pound, ShortTon, SolarMass, Stone,
-        Teragram, Tonne, Yoctogram, Yottagram, Zeptogram, Zettagram,
+        Attogram, Centigram, Decagram, Decigram, Exagram, Femtogram, Gigagram, Gram, Hectogram,
+        Kilogram, MassUnit, Megagram, Microgram, Milligram, Nanogram, Petagram, Picogram, Teragram,
+        Tonne, Yoctogram, Yottagram, Zeptogram, Zettagram,
     };
+    #[cfg(feature = "customary")]
+    pub use qtty_core::units::mass::{Carat, Grain, LongTon, Ounce, Pound, ShortTon, Stone};
+
+    #[cfg(feature = "fundamental-physics")]
+    pub use qtty_core::units::power::ErgPerSecond;
+    #[cfg(feature = "astro")]
+    pub use qtty_core::units::power::SolarLuminosity;
     pub use qtty_core::units::power::{
-        Attowatt, Decawatt, Deciwatt, ErgPerSecond, Exawatt, Femtowatt, Gigawatt, Hectowatt,
-        HorsepowerElectric, HorsepowerMetric, Kilowatt, Megawatt, Microwatt, Milliwatt, Nanowatt,
-        Petawatt, Picowatt, PowerUnit, SolarLuminosity, Terawatt, Watt, Yoctowatt, Yottawatt,
-        Zeptowatt, Zettawatt,
+        Attowatt, Decawatt, Deciwatt, Exawatt, Femtowatt, Gigawatt, Hectowatt, Kilowatt, Megawatt,
+        Microwatt, Milliwatt, Nanowatt, Petawatt, Picowatt, PowerUnit, Terawatt, Watt, Yoctowatt,
+        Yottawatt, Zeptowatt, Zettawatt,
     };
+    #[cfg(feature = "customary")]
+    pub use qtty_core::units::power::{HorsepowerElectric, HorsepowerMetric};
+
     pub use qtty_core::units::time::{
         Attosecond, Centisecond, Century, Day, Decade, Decasecond, Decisecond, Femtosecond,
-        Fortnight, Gigasecond, Hectosecond, Hour, JulianCentury, JulianYear, Kilosecond,
-        Megasecond, Microsecond, Millennium, Millisecond, Minute, Nanosecond, Picosecond, Second,
-        SiderealDay, SiderealYear, SynodicMonth, Terasecond, TimeUnit, Week, Year,
+        Fortnight, Gigasecond, Hectosecond, Hour, Kilosecond, Megasecond, Microsecond, Millennium,
+        Millisecond, Minute, Nanosecond, Picosecond, Second, Terasecond, TimeUnit, Week, Year,
     };
+    #[cfg(feature = "julian-time")]
+    pub use qtty_core::units::time::{JulianCentury, JulianYear};
+    #[cfg(feature = "astro")]
+    pub use qtty_core::units::time::{SiderealDay, SiderealYear, SynodicMonth};
+
     pub use qtty_core::units::volume::{
-        Centiliter, CubicCentimeter, CubicFoot, CubicInch, CubicKilometer, CubicMeter,
-        CubicMillimeter, Deciliter, Liter, Microliter, Milliliter, UsFluidOunce, UsGallon,
-        VolumeUnit,
+        Centiliter, CubicCentimeter, CubicKilometer, CubicMeter, CubicMillimeter, Deciliter, Liter,
+        Microliter, Milliliter, VolumeUnit,
     };
+    #[cfg(feature = "customary")]
+    pub use qtty_core::units::volume::{CubicFoot, CubicInch, UsFluidOunce, UsGallon};
 }
 
 /// Velocity quantities represented as one unit divided by another.
@@ -268,7 +306,9 @@ scalar_aliases::define_scalar_aliases!(f64);
 pub type Unitless<S = f64> = Quantity<unit::Unitless, S>;
 
 pub use qtty_core::units::angular::{DEG, RAD};
-pub use qtty_core::units::length::{AU, KM, LY, M};
+#[cfg(feature = "astro")]
+pub use qtty_core::units::length::{AU, LY};
+pub use qtty_core::units::length::{KM, M};
 pub use qtty_core::units::time::{DAY, SEC};
 
 pub use frequency::Frequency;
