+++ qtty/tests/integration_tests.rs
@@ -37,6 +37,7 @@ fn smoke_test_mass() {
 }
 
 #[test]
+#[cfg(feature = "astro")]
 fn smoke_test_power() {
     let sol = SolarLuminosity::new(1.0);
     let w: Watt = sol.to();
@@ -105,6 +106,7 @@ fn orbital_distance_calculation() {
 }
 
 #[test]
+#[cfg(feature = "astro")]
 fn proxima_centauri_distance() {
     // Proxima Centauri is about 4.24 light years away
     let distance_ly = LightYear::new(4.24);
@@ -144,6 +146,7 @@ fn earth_rotation() {
 }
 
 #[test]
+#[cfg(feature = "astro")]
 fn sun_mass() {
     let sun = SolarMass::new(1.0);
     let kg: Kilogram = sun.to();
@@ -153,6 +156,7 @@ fn sun_mass() {
 }
 
 #[test]
+#[cfg(feature = "astro")]
 fn sun_luminosity() {
     let sun = SolarLuminosity::new(1.0);
     let watts: Watt = sun.to();
@@ -162,6 +166,7 @@ fn sun_luminosity() {
 }
 
 #[test]
+#[cfg(all(feature = "astro", feature = "julian-time"))]
 fn calculate_velocity_from_distance_time() {
     // Light year to km
     let distance = LightYear::new(1.0);
@@ -304,6 +309,7 @@ fn quantity_basic_arithmetic() {
 }
 
 #[test]
+#[cfg(feature = "astro")]
 fn quantity_conversion_chain() {
     // Convert through multiple units
     let au = AstronomicalUnit::new(1.0);
@@ -371,8 +377,11 @@ fn per_unit_division_creates_composite() {
 
 #[test]
 fn unit_constants_have_value_one() {
-    assert_eq!(AU.value(), 1.0);
-    assert_eq!(LY.value(), 1.0);
+    #[cfg(feature = "astro")]
+    {
+        assert_eq!(AU.value(), 1.0);
+        assert_eq!(LY.value(), 1.0);
+    }
     assert_eq!(KM.value(), 1.0);
     assert_eq!(DAY.value(), 1.0);
     assert_eq!(SEC.value(), 1.0);
@@ -382,14 +391,18 @@ fn unit_constants_have_value_one() {
 
 #[test]
 fn constants_can_be_multiplied() {
-    let distance = 4.24 * LY;
-    assert_eq!(distance.value(), 4.24);
+    #[cfg(feature = "astro")]
+    {
+        let distance = 4.24 * LY;
+        assert_eq!(distance.value(), 4.24);
+    }
 
     let time = 365.25 * DAY;
     assert_eq!(time.value(), 365.25);
 }
 
 #[test]
+#[cfg(feature = "astro")]
 fn macro_generated_conversions() {
     // Test conversions that are now generated by impl_unit_conversions! macro
     // These weren't manually implemented before
@@ -413,6 +426,7 @@ fn macro_generated_conversions() {
 }
 
 #[test]
+#[cfg(all(feature = "astro", feature = "navigation"))]
 fn new_angular_units() {
     // Test the new angular units added via impl_unit_conversions! macro
